requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiohttp==3.6.2            # via slackclient
async-timeout==3.0.1      # via aiohttp
attrs==19.3.0             # via aiohttp
chardet==3.0.4            # via aiohttp
idna==2.10                # via yarl
multidict==4.7.6          # via aiohttp, yarl
python-dotenv==0.14.0     # via -r requirements.in
slackclient==2.7.2        # via -r requirements.in
yarl==1.4.2               # via aiohttp
